Move `WorkspaceDetails` and `WorkspaceStateIcon`
