+++ client/web/src/enterprise/batches/batch-spec/execute/workspaces/WorkspaceDetails.mock.ts
@@ -6,7 +6,7 @@ import {
     ChangesetSpecType,
     HiddenBatchSpecWorkspaceFields,
     VisibleBatchSpecWorkspaceFields,
-} from '../../../../graphql-operations'
+} from '../../../../../graphql-operations'
 
 const now = new Date()
