+++ client/web/src/enterprise/batches/batch-spec/execute/workspaces/WorkspaceDetails.module.scss
